backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
qdrant-client
pydantic
httpx